+++ test/REQUIRE
@@ -1,2 +1,1 @@
 StatsBase
-BaseTestNext
